requirements.txt
attrs==23.1.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
jsonschema==4.17.3
pkgutil_resolve_name==1.3.10
PyMySQL==1.1.0
pyrsistent==0.19.3
PyYAML==6.0.1
typing_extensions==4.7.1
zipp==3.15.0
